requirements.txt
streamlit
chromadb
sentence-transformers
llama-index
llama-index-vector-stores-chroma
llama-index-embeddings-huggingface
groq
python-dotenv
plotly
pandas
requests
tqdm